+++ src/rex.restful/src/rex/restful/__init__.py
@@ -4,6 +4,8 @@
 
 
 from .command import *
+from .initialize import *
 from .serializer import *
+from .settings import *
 from .validators import *